tidied up exceptions and deleted locator code
