+++ forest/db/__init__.py
@@ -3,4 +3,3 @@ from .database import *
 from .locate import *
 from .util import *
 from .view import *
-from .exceptions import *
