+++ forest/main.py
@@ -12,7 +12,7 @@ from forest import (
         geo,
         colors,
         db,
-        disk,
+        unified_model,
         navigate,
         parse_args)
 import forest.config as cfg
@@ -126,7 +126,7 @@ def main(argv=None):
                 loader = data.DBLoader(group.label, group.pattern, locator)
             elif group.locator == "file_system":
                 if group.file_type == 'unified_model':
-                    locator = disk.Locator(args.files)
+                    locator = unified_model.Locator(args.files)
                     loader = data.DBLoader(group.label, group.pattern, locator)
                 else:
                     if args.directory is not None:
